refactor: split RiftRoomGenerator into multiple classes
