+++ src/main/java/com/wanderersoftherift/wotr/world/level/levelgen/template/PerimeterGeneratable.java
@@ -2,7 +2,6 @@ package com.wanderersoftherift.wotr.world.level.levelgen.template;
 
 import com.wanderersoftherift.wotr.util.TripleMirror;
 import com.wanderersoftherift.wotr.world.level.levelgen.RiftProcessedRoom;
-import com.wanderersoftherift.wotr.world.level.levelgen.space.CorridorValidator;
 import net.minecraft.core.Vec3i;
 import net.minecraft.world.level.ServerLevelAccessor;
 import net.minecraft.world.level.block.state.BlockState;
@@ -14,7 +13,7 @@ import java.util.List;
 /**
  * places bedrock around the room
  */
-public record PerimeterGeneratable(BlockState perimeterBlock, CorridorValidator validator) implements RiftGeneratable {
+public record PerimeterGeneratable(BlockState perimeterBlock) implements RiftGeneratable {
 
     @Override
     public void processAndPlace(
